clean up vscode-context
