+++ lib/shared/src/chat/recipes/custom-prompt.ts
@@ -1,5 +1,3 @@
-import { URI } from 'vscode-uri'
-
 import { CodebaseContext } from '../../codebase-context'
 import { ContextFile, ContextMessage } from '../../codebase-context/messages'
 import { ActiveTextEditorSelection, Editor } from '../../editor'
@@ -14,21 +12,11 @@ import {
     newInteraction,
     newInteractionWithError,
 } from '../prompts/utils'
-import {
-    getCurrentDirContext,
-    getCurrentFileContextFromEditorSelection,
-    getCurrentFileImportsContext,
-    getDirectoryFileListContext,
-    getEditorDirContext,
-    getEditorOpenTabsContext,
-    getFilePathContext,
-    getPackageJsonContext,
-    getTerminalOutputContext,
-} from '../prompts/vscode-context'
+import { VSCodeEditorContext } from '../prompts/vscode-context/VSCodeEditorContext'
 import { Interaction } from '../transcript/interaction'
 
 import { ChatQuestion } from './chat-question'
-import { getFileExtension, numResults } from './helpers'
+import { numResults } from './helpers'
 import { Recipe, RecipeContext, RecipeID } from './recipe'
 
 /**
@@ -48,8 +36,7 @@ export class CustomPrompt implements Recipe {
     public async getInteraction(commandRunnerID: string, context: RecipeContext): Promise<Interaction | null> {
         const command = context.editor.controllers?.command?.getCommand(commandRunnerID)
         if (!command) {
-            const errorMessage = 'Invalid command -- command not found.'
-            return newInteractionWithError(errorMessage)
+            return createInteractionForError('command')
         }
         const isChatQuestion = command?.slashCommand === '/ask'
 
@@ -67,13 +54,11 @@ export class CustomPrompt implements Recipe {
         const source = getCommandEventSource(command)
 
         if (!promptText) {
-            const errorMessage = 'Please enter a valid prompt for the custom command.'
-            return newInteractionWithError(errorMessage, promptText || '')
+            return createInteractionForError('prompt', promptText)
         }
 
         if (contextConfig?.selection && !selection?.selectedText) {
-            const errorMessage = `__${commandName}__ requires highlighted code. Please select some code in your editor and try again.`
-            return newInteractionWithError(errorMessage, commandName)
+            return createInteractionForError('selection', commandName)
         }
 
         const text = getHumanLLMText(promptText, selection?.fileName)
@@ -88,13 +73,16 @@ export class CustomPrompt implements Recipe {
 
         const truncatedText = truncateText(text, MAX_HUMAN_INPUT_TOKENS)
 
+        const editorContext = new VSCodeEditorContext(context.editor, selection)
+
         // Attach code selection to prompt text if only selection is needed as context
         if (selection && isOnlySelectionRequired(contextConfig)) {
-            const contextMessages = Promise.resolve(getCurrentFileContextFromEditorSelection(selection))
+            const contextMessages = Promise.resolve(editorContext.getCurrentFileContextFromEditorSelection())
             return newInteraction({ text, displayText, contextMessages, source })
         }
 
         const contextMessages = this.getContextMessages(
+            editorContext,
             promptText,
             context.editor,
             context.codebaseContext,
@@ -108,10 +96,11 @@ export class CustomPrompt implements Recipe {
     }
 
     private async getContextMessages(
+        editorContext: VSCodeEditorContext,
         promptText: string,
         editor: Editor,
         codebaseContext: CodebaseContext,
-        promptContext: CodyPromptContext,
+        contextConfig: CodyPromptContext,
         selection?: ActiveTextEditorSelection | null,
         contextFiles?: ContextFile[],
         commandOutput?: string | null
@@ -120,52 +109,53 @@ export class CustomPrompt implements Recipe {
         const workspaceRootUri = editor.getWorkspaceRootUri()
         const isUnitTestRequest = extractTestType(promptText) === 'unit'
 
-        if (promptContext.none) {
+        if (contextConfig.none) {
             return []
         }
 
-        if (promptContext.codebase) {
+        if (contextConfig.codebase) {
             const codebaseMessages = await codebaseContext.getContextMessages(promptText, numResults)
             contextMessages.push(...codebaseMessages)
         }
 
-        if (promptContext.openTabs) {
-            const openTabsMessages = await getEditorOpenTabsContext()
+        if (contextConfig.openTabs) {
+            const openTabsMessages = await editorContext.getEditorOpenTabsContext()
             contextMessages.push(...openTabsMessages)
         }
 
-        if (promptContext.currentDir) {
-            const currentDirMessages = await getCurrentDirContext(isUnitTestRequest)
+        if (contextConfig.currentDir) {
+            const currentDirMessages = await editorContext.getCurrentDirContext(isUnitTestRequest)
             contextMessages.push(...currentDirMessages)
         }
 
-        if (promptContext.directoryPath) {
-            const dirMessages = await getEditorDirContext(promptContext.directoryPath, selection?.fileName)
+        if (contextConfig.directoryPath) {
+            const dirMessages = await editorContext.getEditorDirContext(
+                contextConfig.directoryPath,
+                selection?.fileName
+            )
             contextMessages.push(...dirMessages)
         }
 
-        if (promptContext.filePath) {
-            const fileMessages = await getFilePathContext(promptContext.filePath)
+        if (contextConfig.filePath) {
+            const fileMessages = await editorContext.getFilePathContext(contextConfig.filePath)
             contextMessages.push(...fileMessages)
         }
 
         // Context for unit tests requests
         if (isUnitTestRequest && contextMessages.length === 0) {
             if (selection?.fileName) {
-                const importsContext = await this.getUnitTestContextMessages(selection, workspaceRootUri)
+                const importsContext = await editorContext.getUnitTestContextMessages(selection, workspaceRootUri)
                 contextMessages.push(...importsContext)
             }
         }
 
-        if (promptContext.currentFile || promptContext.selection !== false) {
-            if (selection) {
-                const currentFileMessages = getCurrentFileContextFromEditorSelection(selection)
-                contextMessages.push(...currentFileMessages)
-            }
+        if (contextConfig.currentFile || contextConfig.selection !== false) {
+            const currentFileMessages = editorContext.getCurrentFileContextFromEditorSelection()
+            contextMessages.push(...currentFileMessages)
         }
 
-        if (promptContext.command && commandOutput) {
-            const outputMessages = getTerminalOutputContext(commandOutput)
+        if (contextConfig.command && commandOutput) {
+            const outputMessages = editorContext.getTerminalOutputContext(commandOutput)
             contextMessages.push(...outputMessages)
         }
 
@@ -178,28 +168,18 @@ export class CustomPrompt implements Recipe {
         const maxResults = Math.floor((NUM_CODE_RESULTS + NUM_TEXT_RESULTS) / 2) * 2
         return contextMessages.slice(-maxResults * 2)
     }
+}
 
-    private async getUnitTestContextMessages(
-        selection: ActiveTextEditorSelection,
-        workspaceRootUri?: URI | null
-    ): Promise<ContextMessage[]> {
-        const contextMessages: ContextMessage[] = []
-
-        if (workspaceRootUri) {
-            const rootFileNames = await getDirectoryFileListContext(workspaceRootUri, true)
-            contextMessages.push(...rootFileNames)
-        }
-        // Add package.json content only if files matches the ts/js extension regex
-        if (selection?.fileName && getFileExtension(selection?.fileName).match(/ts|js/)) {
-            const packageJson = await getPackageJsonContext(selection?.fileName)
-            contextMessages.push(...packageJson)
-        }
-        // Try adding import statements from current file as context
-        if (selection?.fileName) {
-            const importsContext = await getCurrentFileImportsContext()
-            contextMessages.push(...importsContext)
-        }
-
-        return contextMessages
+function createInteractionForError(errorType: 'command' | 'prompt' | 'selection', args?: string): Promise<Interaction> {
+    switch (errorType) {
+        case 'command':
+            return newInteractionWithError('Invalid command -- command not found.')
+        case 'prompt':
+            return newInteractionWithError('Please enter a valid prompt for the custom command.', args || '')
+        case 'selection':
+            return newInteractionWithError(
+                `__${args}__ requires highlighted code. Please select some code in your editor and try again.`,
+                args
+            )
     }
 }
